merging easter-fix into migrate-oauth-to-github-app for full testing on dev* servers

+++ deploy/restart-apache.sh
@@ -7,6 +7,10 @@ OPENTREE_HOST=$2
 CERTIFICATE_FILE=$3
 CERTIFICATE_KEY_FILE=$4
 OTINDEX_BASE_URL=$5
+PHYLAPI_BASE_URL=$6
+update_apache_config=$7
+installing_web2py=$8
+
 
 OPENTREE_HOME=$(bash <<< "echo ~$OPENTREE_USER")
 
@@ -14,42 +18,49 @@ if apt-cache policy apache2 | egrep -q "Installed: 2.2"; then
 
 echo "This project requires apache 2.4 or higher! Please upgrade apache and try again."
 exit 1
-
-else
+fi
+#else
 
 # Apache 2.4+
 
-if [ ! -r /etc/apache2/sites-available/opentree.conf ] || \
-   ! cmp -s "$OPENTREE_HOME/setup/opentree.conf" /etc/apache2/sites-available/opentree; then
-    echo "Installing opentree vhost config"
-    sudo cp -p "$OPENTREE_HOME/setup/opentree.conf" /etc/apache2/sites-available/ || \
-      "opentree.conf install failed"
-fi
+if [ "$update_apache_config" = "yes" ]; then
 
-if [ ! -r /etc/apache2/sites-available/opentree-ssl.conf ] || \
-   ! cmp -s "$OPENTREE_HOME/setup/opentree-ssl.conf" /etc/apache2/sites-available/opentree-ssl.conf; then
-    echo "Installing opentree ssl vhost config"
-    # This will overwrite the letsencrypt configuration
-    sudo cp -p "$OPENTREE_HOME/setup/opentree-ssl.conf" /etc/apache2/sites-available/ || \
-      "opentree-ssl.conf install failed"
-    sudo sed -i -e s/SERVERNAME_REPLACEME/$OPENTREE_HOST/g \
-      /etc/apache2/sites-available/opentree-ssl.conf  || "Edit hostname in opentree-ssl failed"
-    sudo sed -i -e s+CERTIFICATE_FILE+$CERTIFICATE_FILE+ \
-      /etc/apache2/sites-available/opentree-ssl.conf  || "Edit cert file in opentree-ssl failed"
-    sudo sed -i -e s+CERTIFICATE_KEY_FILE+$CERTIFICATE_KEY_FILE+ \
-      /etc/apache2/sites-available/opentree-ssl.conf  || "Edit cert key file in opentree-ssl failed"
-fi
+  echo "Updating apache configuration..."   # TODO
 
-TMP=/tmp/$$.tmp
-sed -e s+/home/opentree+$OPENTREE_HOME+ <"$OPENTREE_HOME/setup/opentree-shared.conf" >$TMP
-sed -i -e "s+OTINDEX_BASE_URL+$OTINDEX_BASE_URL+" $TMP
-if [ ! -r /etc/apache2/opentree-shared.conf ] || \
-   ! cmp -s $TMP /etc/apache2/opentree-shared.conf; then
-    echo "Installing opentree vhosts shared config"
-    sudo cp -p $TMP /etc/apache2/opentree-shared.conf || "Sudo failed"
-    # Access otindex search from shared server-config variable
-fi
-rm $TMP
+  if [ ! -r /etc/apache2/sites-available/opentree.conf ] || \
+     ! cmp -s "$OPENTREE_HOME/setup/opentree.conf" /etc/apache2/sites-available/opentree; then
+      echo "Installing opentree vhost config"
+      sudo cp -p "$OPENTREE_HOME/setup/opentree.conf" /etc/apache2/sites-available/ || \
+        "opentree.conf install failed"
+  fi
+  
+  if [ ! -r /etc/apache2/sites-available/opentree-ssl.conf ] || \
+     ! cmp -s "$OPENTREE_HOME/setup/opentree-ssl.conf" /etc/apache2/sites-available/opentree-ssl.conf; then
+      echo "Installing opentree ssl vhost config"
+      # This will overwrite the letsencrypt configuration
+      sudo cp -p "$OPENTREE_HOME/setup/opentree-ssl.conf" /etc/apache2/sites-available/ || \
+        "opentree-ssl.conf install failed"
+      sudo sed -i -e s/SERVERNAME_REPLACEME/$OPENTREE_HOST/g \
+        /etc/apache2/sites-available/opentree-ssl.conf  || "Edit hostname in opentree-ssl failed"
+      sudo sed -i -e s+CERTIFICATE_FILE+$CERTIFICATE_FILE+ \
+        /etc/apache2/sites-available/opentree-ssl.conf  || "Edit cert file in opentree-ssl failed"
+      sudo sed -i -e s+CERTIFICATE_KEY_FILE+$CERTIFICATE_KEY_FILE+ \
+        /etc/apache2/sites-available/opentree-ssl.conf  || "Edit cert key file in opentree-ssl failed"
+  fi
+  
+  TMP=/tmp/$$.tmp
+  sed -e s+/home/opentree+$OPENTREE_HOME+ <"$OPENTREE_HOME/setup/opentree-shared.conf" >$TMP
+  sed -i -e "s+OTINDEX_BASE_URL+$OTINDEX_BASE_URL+" $TMP
+  sed -i -e "s+PHYLAPI_BASE_URL+$PHYLAPI_BASE_URL+" $TMP
+  
+  if [ ! -r /etc/apache2/opentree-shared.conf ] || \
+     ! cmp -s $TMP /etc/apache2/opentree-shared.conf; then
+      echo "Installing opentree vhosts shared config"
+     sudo cp -p $TMP /etc/apache2/opentree-shared.conf || "Sudo failed"
+     # Access otindex search from shared server-config variable
+  fi
+  rm $TMP
+
 fi
 
 echo "Restarting apache httpd..."
@@ -59,7 +70,7 @@ sudo apache2ctl graceful || echo "apache2ctl failed"
 # so for now I'm going to disable this code.  See 
 # https://github.com/OpenTreeOfLife/opentree/issues/845
 
-if false; then
+if [ "$installing_web2py" = "yes" ]; then
   echo "(Re)starting web2py session sweeper..."
   # The sessions2trash.py utility script runs in the background, deleting expired
   # sessions every 5 minutes. See documentation at
